cart

+++ frontend.sh
@@ -9,18 +9,17 @@ N="\e[0m"
 LOGS_FOLDER="/var/log/shell-roboshop"
 SCRIPT_NAME=$( echo $0 | cut -d "." -f1 )
 SCRIPT_DIR=$PWD
-MONGODB_HOST=mongodb.daws86.space
-LOG_FILE="$LOGS_FOLDER/$SCRIPT_NAME.log" # /var/log/shell-script/16-logs.log
+LOG_FILE="$LOGS_FOLDER/$SCRIPT_NAME.log"
 
 mkdir -p $LOGS_FOLDER
 echo "Script started executed at: $(date)" | tee -a $LOG_FILE
 
 if [ $USERID -ne 0 ]; then
-    echo "ERROR:: Please run this script with root privelege"
-    exit 1 # failure is other than 0
+    echo "ERROR:: Please run this script with root privilege"
+    exit 1
 fi
 
-VALIDATE(){ # functions receive inputs through args just like shell script args
+VALIDATE() {
     if [ $1 -ne 0 ]; then
         echo -e "$2 ... $R FAILURE $N" | tee -a $LOG_FILE
         exit 1
@@ -29,26 +28,33 @@ VALIDATE(){ # functions receive inputs through args just like shell script args
     fi
 }
 
+# Install Nginx
 dnf module disable nginx -y &>>$LOG_FILE
 dnf module enable nginx:1.24 -y &>>$LOG_FILE
 dnf install nginx -y &>>$LOG_FILE
 VALIDATE $? "Installing Nginx"
 
-systemctl enable nginx  &>>$LOG_FILE
-systemctl start nginx 
-VALIDATE $? "Starting Nginx"
-
-rm -rf /usr/share/nginx/html/* 
+# Remove default content and deploy frontend
+rm -rf /usr/share/nginx/html/*
 curl -o /tmp/frontend.zip https://roboshop-artifacts.s3.amazonaws.com/frontend-v3.zip &>>$LOG_FILE
-cd /usr/share/nginx/html 
+cd /usr/share/nginx/html
 unzip /tmp/frontend.zip &>>$LOG_FILE
-VALIDATE $? "Downloading frontend"
+VALIDATE $? "Downloading & Extracting frontend"
 
-rm -rf /etc/nginx/nginx.conf
-cp $SCRIPT_DIR/nginx.conf /etc/nginx/nginx.conf
-VALIDATE $? "Copying nginx.conf"
+# Deploy custom nginx.conf
+if [ -f $SCRIPT_DIR/nginx.conf ]; then
+    cp $SCRIPT_DIR/nginx.conf /etc/nginx/nginx.conf
+    VALIDATE $? "Copying nginx.conf"
+else
+    echo -e "Custom nginx.conf missing ... $R FAILURE $N" | tee -a $LOG_FILE
+    exit 1
+fi
 
+# Validate nginx config before restart
+nginx -t &>>$LOG_FILE
+VALIDATE $? "Validating Nginx config"
 
-
-systemctl restart nginx 
-VALIDATE $? "Restarting Nginx"
+# Enable and start nginx
+systemctl enable nginx &>>$LOG_FILE
+systemctl restart nginx &>>$LOG_FILE
+VALIDATE $? "Starting Nginx"
